+++ docs/changelog.md
@@ -4,3 +4,10 @@
 - Initial project commit
 - Initial project structure setup 
 - Added basic OpenGL window and context creation using GLFW
+
+## 2026-01-25
+- Implemented input system handling for keyboard and mouse
+- Added a basic Shader class for loading and compiling GLSL shaders into a shader program
+- Added GLM for mathematical operations
+- Integrated TriangleRenderer into the engine loop for testing the input system and shader class
+- Added dynamic triangle creation using the space key in TriangleRenderer